# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v1/registration_questions:
    get:
      tags:
        - RegistrationQuestions
      summary: List RegistrationQuestions
      description: Get a paginated list of RegistrationQuestions with optional filtering and sorting
      operationId: listRegistrationQuestions
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegistrationQuestionList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - RegistrationQuestions
      summary: Create RegistrationQuestion
      description: Create a new RegistrationQuestion
      operationId: createRegistrationQuestion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRegistrationQuestionRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegistrationQuestion'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/tickets:
    get:
      tags:
        - Tickets
      summary: List Tickets
      description: Get a paginated list of Tickets with optional filtering and sorting
      operationId: listTickets
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Tickets
      summary: Create Ticket
      description: Create a new Ticket
      operationId: createTicket
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTicketRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ticket'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    RegistrationQuestion:
      type: object
      required:
        - id
        - question_id
        - sequence
        - metadata
      properties:
        id:
          type: string
          format: uuid
        event_id:
          nullable: true
          type: string
          format: uuid
        event_type_id:
          nullable: true
          type: string
          format: uuid
        question_id:
          type: string
          format: uuid
        sequence:
          type: integer
          format: int32
        metadata:
          type: object
...
    CreateRegistrationQuestionRequest:
      type: object
      required:
        - question_id
        - sequence
        - metadata
      properties:
        event_id:
          nullable: true
          type: string
          format: uuid
        event_type_id:
          nullable: true
          type: string
          format: uuid
        question_id:
          type: string
          format: uuid
        sequence:
          type: integer
          format: int32
        metadata:
          type: object
...
    RegistrationQuestionList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/RegistrationQuestion'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    Ticket:
      type: object
      required:
        - id
        - event_id
        - name
        - seats_max
        - seats_max_per_order
        - metadata
      properties:
        id:
          type: string
          format: uuid
        event_id:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 200
        seats_max:
          type: integer
          format: int32
        seats_max_per_order:
          type: integer
          format: int32
        start_sale_datetime:
          nullable: true
          type: string
          format: date-time
        end_sale_datetime:
          nullable: true
          type: string
          format: date-time
        metadata:
          type: object
...
    CreateTicketRequest:
      type: object
      required:
        - event_id
        - name
        - seats_max
        - seats_max_per_order
        - metadata
      properties:
        event_id:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 200
        seats_max:
          type: integer
          format: int32
        seats_max_per_order:
          type: integer
          format: int32
        start_sale_datetime:
          nullable: true
          type: string
          format: date-time
        end_sale_datetime:
          nullable: true
          type: string
          format: date-time
        metadata:
          type: object
...
    TicketList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Ticket'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []