# src/presentation/http/ticket_handler.rs
//! Ticket REST handlers
//!
//! Generated by metaphor-schema. Do not edit manually.
//!
//! Uses Axum and backbone-core's BackboneCrudHandler for all 12 CRUD endpoints.

use std::collections::HashMap;
use std::sync::Arc;

use axum::Router;
use serde::{Deserialize, Serialize};
use uuid::Uuid;
use chrono::{DateTime, Utc};

// Backbone framework imports
use backbone_core::http::BackboneCrudHandler;

// Auth integration (optional)
#[cfg(feature = "auth")]
use backbone_auth::middleware::AuthContext;
#[cfg(feature = "auth")]
use backbone_auth::AuthMiddleware;

// Domain imports
use crate::domain::entity::*;
use crate::application::service::{TicketService, ServiceError};

// DTO imports
use crate::presentation::dto::{CreateTicketDto, UpdateTicketDto, PatchTicketDto, TicketResponseDto};


/// Application error type
#[derive(Debug, thiserror::Error)]
pub enum TicketError {
    #[error("Not found: {0}")]
    NotFound(String),
    #[error("Validation error: {0}")]
    Validation(String),
    #[error("Database error: {0}")]
    Database(String),
    #[error("Internal error: {0}")]
    Internal(String),
}

impl From<ServiceError> for TicketError {
    fn from(err: ServiceError) -> Self {
        match err {
            ServiceError::NotFound => Self::NotFound(err.to_string()),
            ServiceError::Validation(ref msg) => Self::Validation(msg.clone()),
            ServiceError::AlreadyExists(ref msg) => Self::Validation(msg.clone()),
            ServiceError::Repository(ref e) => Self::Database(e.to_string()),
            ServiceError::Internal(ref msg) => Self::Internal(msg.clone()),
        }
    }
}

impl axum::response::IntoResponse for TicketError {
    fn into_response(self) -> axum::response::Response {
        use axum::http::StatusCode;
        use axum::Json;

        let (status, code) = match &self {
            Self::NotFound(_) => (StatusCode::NOT_FOUND, "TICKET_NOT_FOUND"),
            Self::Validation(_) => (StatusCode::BAD_REQUEST, "TICKET_VALIDATION_ERROR"),
            Self::Database(_) => (StatusCode::INTERNAL_SERVER_ERROR, "TICKET_DATABASE_ERROR"),
            Self::Internal(_) => (StatusCode::INTERNAL_SERVER_ERROR, "TICKET_INTERNAL_ERROR"),
        };

        let body = serde_json::json!({
            "success": false,
            "error": code,
            "message": self.to_string(),
        });

        (status, Json(body)).into_response()
    }
}

// =============================================================================
// Route Configuration
// =============================================================================

/// Create Axum router with all 16 Backbone endpoints for Ticket.
///
/// # Routes
///
/// | Method | Path | Description |
/// |--------|------|-------------|
/// | GET | /tickets | List with pagination |
/// | POST | /tickets | Create new |
/// | GET | /tickets/:id | Get by ID |
/// | PUT | /tickets/:id | Full update |
/// | PATCH | /tickets/:id | Partial update |
/// | DELETE | /tickets/:id | Soft delete |
/// | POST | /tickets/bulk | Bulk create |
/// | POST | /tickets/upsert | Upsert |
/// | GET | /tickets/trash | List deleted |
/// | POST | /tickets/:id/restore | Restore |
/// | DELETE | /tickets/empty | Empty trash |
/// | GET | /tickets/:id/deleted | Get deleted by ID |
/// | DELETE | /tickets/trash/:id | Permanent delete from trash |
/// | GET | /tickets/count | Count active entities |
/// | GET | /tickets/trash/count | Count deleted entities |
///
/// # Example
///
/// ```text
/// let service = Arc::new(TicketService::with_repository(repository));
/// let router = create_ticket_routes(service);
/// ```
pub fn create_ticket_routes(service: Arc<TicketService>) -> Router {
    BackboneCrudHandler::<TicketService, Ticket, CreateTicketDto, UpdateTicketDto, TicketResponseDto>::routes(
        service,
        "/tickets",
    )
}

/// Create Axum router with only the read (GET) endpoints for Ticket.
///
/// Safe for public, unauthenticated exposure (e.g., reference data).
/// Mutations must be served separately via `create_ticket_write_routes`,
/// typically wrapped in an auth middleware layer.
pub fn create_ticket_read_routes(service: Arc<TicketService>) -> Router {
    BackboneCrudHandler::<TicketService, Ticket, CreateTicketDto, UpdateTicketDto, TicketResponseDto>::read_routes(
        service,
        "/tickets",
    )
}

/// Create Axum router with only the write (mutation) endpoints for Ticket.
///
/// These routes must NOT be publicly exposed. Wrap them with an auth
/// middleware before nesting into the application router.
///
/// # This is unguarded generic CRUD, not a validated write path
///
/// These are plain create/update/patch/delete mutations over the entity row —
/// they bypass all business invariants. If the module exposes a validated write
/// service (e.g. a command router over its domain engine), serve THAT instead
/// for any mutation that must respect domain rules.
pub fn create_ticket_write_routes(service: Arc<TicketService>) -> Router {
    BackboneCrudHandler::<TicketService, Ticket, CreateTicketDto, UpdateTicketDto, TicketResponseDto>::write_routes(
        service,
        "/tickets",
    )
}

/// Create authenticated routes with auth middleware.
///
/// Requires the `auth` feature flag. The `AuthMiddleware` implementation
/// is responsible for extracting and validating tokens, then providing
/// an `AuthContext` via request extensions.
#[cfg(feature = "auth")]
pub fn create_protected_ticket_routes<A: AuthMiddleware + Send + Sync + 'static>(
    service: Arc<TicketService>,
    auth: Arc<A>,
) -> Router {
    use axum::middleware;
    use axum::response::IntoResponse;

    let auth_layer = auth.clone();
    create_ticket_routes(service)
        .layer(middleware::from_fn(move |mut req: axum::extract::Request, next: axum::middleware::Next| {
            let auth = auth_layer.clone();
            async move {
                let token = req.headers()
                    .get(axum::http::header::AUTHORIZATION)
                    .and_then(|h| h.to_str().ok())
                    .and_then(|raw| raw.strip_prefix("Bearer ").or_else(|| raw.strip_prefix("bearer ")))
                    .unwrap_or("");
                match auth.authenticate(token).await {
                    Ok(ctx) => {
                        req.extensions_mut().insert(ctx);
                        next.run(req).await
                    }
                    Err(_) => {
                        (axum::http::StatusCode::UNAUTHORIZED,
                         axum::Json(serde_json::json!({
                             "success": false,
                             "error": "unauthorized",
                             "message": "Authentication required"
                         }))
                        ).into_response()
                    }
                }
            }
        }))
}